WIP: add memoized masonry lib and comps
